+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "agenthub-codex-acp"]
+members = [".", "agenthub-codex-acp", "crates/agenthub-acp-core", "crates/agenthub-acp"]
 
 [patch.crates-io]
 tokio-tungstenite = { git = "https://github.com/JakkuSakura/tokio-tungstenite", branch = "proxy-support" }
@@ -40,4 +40,5 @@ jwt-simple = "0.11"
 mime_guess = "2"
 rust-embed = "8"
 async-trait = "0.1"
+agenthub-acp = { path = "crates/agenthub-acp" }
 agent-client-protocol = { version = "0.9.4", features = ["unstable"] }
