+++ crates/agenthub-acp-core/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "agenthub-acp-core"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+agent-client-protocol = { version = "0.9.4", features = ["unstable"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
