+++ crates/agenthub-acp/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "agenthub-acp"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+agenthub-acp-core = { path = "../agenthub-acp-core" }
+agent-client-protocol = { version = "0.9.4", features = ["unstable"] }
+anyhow = "1"
+async-trait = "0.1"
+chrono = { version = "0.4", features = ["clock"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sqlx = { version = "0.8.6", features = ["runtime-tokio", "sqlite", "chrono", "uuid"] }
+tokio = { version = "1", features = ["macros", "rt-multi-thread", "process", "sync", "time", "io-util"] }
+tokio-util = { version = "0.7", features = ["compat"] }
+tracing = "0.1"
+uuid = { version = "1", features = ["v4", "v7", "serde"] }
